requirements.txt
yfinance
croniter
python-telegram-bot
python-decouple




